ref: Move blocks, timelines, and trials into "experiment" folder
