+++ src/experiment/blocks/startBlock.js
@@ -1,4 +1,4 @@
-import { config } from "../config/main";
+import { config } from "../../config/main";
 
 import { buildCameraStartTrial } from "../trials/camera";
 import { enterFullscreenTrial } from "../trials/fullscreen";
