+++ src/experiment/trials/conclusion.js
@@ -1,7 +1,7 @@
 import htmlKeyboardResponse from "@jspsych/plugin-html-keyboard-response";
 
-import { LANGUAGE } from "../config/main";
-import { h1 } from "../lib/markup/tags";
+import { LANGUAGE } from "../../config/main";
+import { h1 } from "../../lib/markup/tags";
 
 /** Trial that displays a completion message for 5 seconds */
 const conclusionTrial = {
