+++ src/experiment/trials/fixation.js
@@ -1,8 +1,8 @@
 import htmlKeyboardResponse from "@jspsych/plugin-html-keyboard-response";
 
-import { SETTINGS, config, eventCodes } from "../config/main";
-import { pdSpotEncode, photodiodeGhostBox } from "../lib/markup/photodiode";
-import { div } from "../lib/markup/tags";
+import { SETTINGS, config, eventCodes } from "../../config/main";
+import { pdSpotEncode, photodiodeGhostBox } from "../../lib/markup/photodiode";
+import { div } from "../../lib/markup/tags";
 
 /**
  * Builds a trial with a fixation dot and optional photodiode box.
