+++ src/experiment/trials/fullscreen.js
