+++ src/experiment/trials/introduction.js
@@ -1,7 +1,7 @@
 import htmlKeyboardResponse from "@jspsych/plugin-html-keyboard-response";
 
-import { LANGUAGE } from "../config/main";
-import { div, h1 } from "../lib/markup/tags";
+import { LANGUAGE } from "../../config/main";
+import { div, h1 } from "../../lib/markup/tags";
 
 /** Task that displays a welcome message with the photodiode ghost box */
 const introductionTrial = {
