+++ src/experiment/trials/name.js
@@ -1,7 +1,7 @@
 import htmlKeyboardResponse from "@jspsych/plugin-html-keyboard-response";
 
-import { LANGUAGE } from "../config/main";
-import { h1 } from "../lib/markup/tags";
+import { LANGUAGE } from "../../config/main";
+import { h1 } from "../../lib/markup/tags";
 
 /** Task that displays the name of the experiment */
 const nameTrial = {
