+++ src/experiment/trials/survey.js
@@ -1,6 +1,6 @@
 import jsPsychSurveyPlugin from "@jspsych/plugin-survey";
 
-import { LANGUAGE } from "../config/main";
+import { LANGUAGE } from "../../config/main";
 
 // The survey plugin includes additional styling
 import "@jspsych/plugin-survey/css/survey.css";
